requirements.txt
# ML
scikit-learn>=1.6
xgboost>=3.0

# NLP / embeddings
sentence-transformers>=5.0

# Data & config
numpy>=1.26
pandas>=2.3
PyYAML>=6.0

# Explainability & visualization
shap>=0.46
matplotlib>=3.10

# Utilities
joblib>=1.4
tqdm>=4.67
